+++ src/main/kotlin/com/cquilez/pitesthelper/listeners/MyFrameStateListener.kt
@@ -1,4 +1,4 @@
-package com.github.carmeloquilez.pitestmavenhelper.listeners
+package com.cquilez.pitesthelper.listeners
 
 import com.intellij.ide.FrameStateListener
 import com.intellij.openapi.diagnostic.thisLogger